requirements.txt
fastapi
progressbar2
xlrd
